Save Waypoints, Fix Start Route, Add Persistent Local Data
Made it so a GPX file saves in local storage, and then updates on each new app load with any new waypoints made. Fixed an issue where if you were close enough to a route to actually start the cycle, it just told you the route had started without actually starting it.