Fix premature route stoppage error
Fixed an issue where stopping the route before valid info was saved would still show the GPX name change modal, which attempts to change the name of a GPX fie that no longer exists.

Fixed an issue where starting a route only checked if there was a currentGPXPath, not if a file actually existed there.